+++ src/main/java/amlsim/model/aml/FanInTypology.java
@@ -22,7 +22,7 @@ public class FanInTypology extends AMLTypology {
     private static final int FIXED_INTERVAL = 2;
     private static final int RANDOM_RANGE = 3;
 
-    FanInTypology(float minAmount, float maxAmount, int start, int end){
+    FanInTypology(double minAmount, double maxAmount, int start, int end){
         super(minAmount, maxAmount, start, end);
     }
 
@@ -79,7 +79,7 @@ public class FanInTypology extends AMLTypology {
     public void sendTransactions(long step, Account acct){
         long alertID = alert.getAlertID();
         boolean isSAR = alert.isSAR();
-        float amount = getRandomAmount();
+        double amount = getRandomAmount();
 
         for(int i = 0; i< origList.size(); i++){
             if(steps[i] == step){
