+++ src/main/java/amlsim/model/aml/GatherScatterTypology.java
@@ -14,10 +14,10 @@ public class GatherScatterTypology extends AMLTypology {
     private long[] gatherSteps;
     private long[] scatterSteps;
     private long middleStep;
-    private float totalReceivedAmount = 0.0F;
-    private float scatterAmount = 0.0F;  // Scatter transaction amount will be defined after gather transactions
+    private double totalReceivedAmount = 0.0;
+    private double scatterAmount = 0.0;  // Scatter transaction amount will be defined after gather transactions
 
-    GatherScatterTypology(float minAmount, float maxAmount, int startStep, int endStep) {
+    GatherScatterTypology(double minAmount, double maxAmount, int startStep, int endStep) {
         super(minAmount, maxAmount, startStep, endStep);
     }
 
@@ -78,7 +78,7 @@ public class GatherScatterTypology extends AMLTypology {
                 if(gatherSteps[i] == step){
                     Account orig = origAccts.get(i);
                     Account bene = alert.getMainAccount();
-                    float amount = getRandomAmount();
+                    double amount = getRandomAmount();
                     makeTransaction(step, amount, orig, bene, isSAR, alertID);
                     totalReceivedAmount += amount;
                 }
@@ -93,7 +93,7 @@ public class GatherScatterTypology extends AMLTypology {
             }
         }
         if(step == middleStep){  // Define the amount of scatter transactions
-            float margin = totalReceivedAmount * marginRatio;
+            double margin = totalReceivedAmount * marginRatio;
             scatterAmount = (totalReceivedAmount - margin) / numScatters;
         }
     }
